pyproject.toml
[project]
name = "hystron-node"
version = "0.1.0"
description = "Xray-core node agent for the Hystron panel"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.64",
    "grpcio-tools>=1.64",
    "protobuf>=5.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
indent-width = 4
respect-gitignore = true
exclude = ["app/db/migrations"]
fix = true

[tool.ruff.lint]
fixable = ["ALL"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[dependency-groups]
dev = [
    "ruff>=0.11",
]
